requirements.txt
et_xmlfile==2.0.0
greenlet==3.2.4
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
